Move RDB test to test_rdb.py
